+++ cmd/gotoi/main.go
@@ -10,6 +10,8 @@ import (
 	"os/exec"
 	"strconv"
 	"strings"
+
+	"github.com/gregoryv/vt100"
 )
 
 func main() {
@@ -52,16 +54,39 @@ func main() {
 		switch {
 		case len(args) == 0:
 			if part[0] != lastFile {
-				if i > 1 {
-					fmt.Println()
-				}
 				lastFile = part[0]
-				fmt.Println(lastFile)
+				fmt.Print(fg.Cyan.String() +attr.Dim.String() + strings.TrimSpace(lastFile), attr.Reset, "\n")
 			}
-			fmt.Printf("%v %s\n", i, part[4])
+			fmt.Printf("%v %s\n", i, paint(part[4]))
 
 		case index[i]:
 			exec.Command("emacsclient", "-n", "+"+part[3], part[0]).Run()
 		}
 	}
 }
+
+func paint(v string) string {
+	first := v
+	i := strings.Index(v, " ")
+	if i > 0 {
+		first = first[:i]
+	} else {
+		i = 0
+	}
+	switch first {
+	case "import":
+		return fg.Magenta.String() + first + attr.Reset.String()
+
+	case "var", "const", "func", "type", "package":
+		return fg.Magenta.String() + first + attr.Reset.String() + v[i:]
+
+	case "//":
+		return fg.Green.String() + attr.Dim.String() + v + attr.Reset.String()
+	}
+	return v
+}
+
+var (
+	fg   = vt100.ForegroundColors()
+	attr = vt100.Attributes()
+)